+++ src/data/index.ts
@@ -4,21 +4,25 @@ import { initData } from "./main"
 import { ParkAndScenarioData as PASD } from "./park_and_scenario"
 import { PlayerData as PD } from "./player"
 import { RideData as RD } from "./rides"
-import { StallsAndFacilitiesData as SAFD } from "./stalls_and_facilities"
+import { SFData as SAFD } from "./stalls_and_facilities"
+import { Options as OPT } from "./options"
+import { FinanceData } from "./finance"
 
 namespace Data {
   export import GuestData = GD
   export import ParkAndScenarioData = PASD
   export import PlayerData = PD
   export import RideData = RD
-  export import StallsAndFacilitiesData = SAFD
+  export import SFData = SAFD
+  export import Options = OPT
 
   export function updateAll(): void {
     PlayerData.update()
     GuestData.update()
     ParkAndScenarioData.update()
-    StallsAndFacilitiesData.update()
+    SFData.update()
     RideData.update()
+    FinanceData.update()
   }
 
   export function init() {
@@ -26,8 +30,9 @@ namespace Data {
     PlayerData.init()
     GuestData.init()
     ParkAndScenarioData.init()
-    StallsAndFacilitiesData.init()
+    SFData.init()
     RideData.init()
+    FinanceData.init()
     interval.initCounter()
   }
 }