+++ src/ui/main.ts
@@ -15,9 +15,9 @@ import { baseData } from "../data/main"
 import { getCurrencyUnit } from "../data/currency"
 import interval from "../utils/interval"
 import Sprites from "./custom/sprites"
-import Options from "../data/options"
 import Data from "../data/index"
 import { progressBar } from "./custom/progress_bar"
+import { GuestData } from "../data/guest"
 
 /**
  * Whether the window is open.
@@ -27,9 +27,11 @@ let isOpen: boolean = false
 let manualIndicatorLit: boolean = false
 
 function initMainMenu(): void {
-  if (typeof ui !== "undefined") {
-    ui.registerMenuItem(language.ui.main.title, menu)
-  }
+  ui.registerMenuItem(language.ui.main.title, menu)
+}
+
+function openMainMenu(): void {
+  menu()
 }
 
 /**
@@ -45,15 +47,15 @@ function createIndicator(pos: number): WidgetCreator<FlexiblePosition> {
       baseData.local.options.countdown_progress.store,
       (v1, v2) => {
         switch (v1) {
-          case Options.UpdateStatus.RUNNING:
+          case Data.Options.UpdateStatus.RUNNING:
             if (pos <= v2 * (10 / baseData.global.update_ratio.get()))
               return Sprites.INDICATOR_RUNNING_LIT
             return Sprites.INDICATOR_RUNNING_UNLIT
-          case Options.UpdateStatus.MANUAL:
+          case Data.Options.UpdateStatus.MANUAL:
             return manualIndicatorLit
               ? Sprites.INDICATOR_MANUAL_LIT
               : Sprites.INDICATOR_MANUAL_UNLIT
-          case Options.UpdateStatus.PAUSED:
+          case Data.Options.UpdateStatus.PAUSED:
             return Sprites.INDICATOR_PAUSED_LIT
           default:
             return -1
@@ -97,7 +99,7 @@ function menu(): void {
           // Player
           groupbox({
             text: language.ui.main.groupbox.player.title,
-            // width: "30%",
+            height: "50%",
             content: [
               label({
                 text:
@@ -214,7 +216,7 @@ function menu(): void {
           // Guest
           groupbox({
             text: language.ui.main.groupbox.guest.title,
-            // width: "40%",
+            height: "50%",
             content: [
               label({
                 text: compute(
@@ -273,9 +275,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_happiness_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_happiness_ave +
                               value.toString()
@@ -290,7 +292,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -298,14 +300,15 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_happiness_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_HAPPINESS
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_happiness_ave.store,
                       (value) => {
-                        if (value / 255 < 0.5) return Colour.BrightRed
+                        if (value < GuestData.HAPINESS_WARNING_THRESHOLD)
+                          return Colour.BrightRed
                         return Colour.BrightGreen
                       }
                     )
@@ -321,9 +324,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_energy_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_energy_ave +
                               value.toString()
@@ -338,7 +341,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -346,14 +349,15 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_energy_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_ENERGY
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_energy_ave.store,
                       (value) => {
-                        if (value / 255 < 0.5) return Colour.BrightRed
+                        if (value < GuestData.ENERGY_WARNING_THRESHOLD)
+                          return Colour.BrightRed
                         return Colour.BrightGreen
                       }
                     )
@@ -369,9 +373,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_nausea_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_nausea_ave +
                               value.toString()
@@ -386,7 +390,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -394,15 +398,16 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_nausea_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_NAUSEA
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_nausea_ave.store,
                       (value) => {
-                        if (value / 255 < 0.5) return Colour.BrightGreen
-                        return Colour.BrightRed
+                        if (value > GuestData.NAUSEA_WARNING_THRESHOLD)
+                          return Colour.BrightRed
+                        return Colour.BrightGreen
                       }
                     )
                   })
@@ -417,9 +422,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_hunger_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_hunger_ave +
                               value.toString()
@@ -434,7 +439,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -442,14 +447,15 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_hunger_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_HUNGER
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_hunger_ave.store,
                       (value) => {
-                        if (value / 255 > 0.5) return Colour.BrightRed
+                        if (value > GuestData.HUNGER_WARNING_THRESHOLD)
+                          return Colour.BrightRed
                         return Colour.BrightGreen
                       }
                     )
@@ -465,9 +471,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_thirst_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_thirst_ave +
                               value.toString()
@@ -482,7 +488,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -490,14 +496,15 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_thirst_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_THIRST
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_thirst_ave.store,
                       (value) => {
-                        if (value / 255 > 0.5) return Colour.BrightRed
+                        if (value > GuestData.THIRST_WARNING_THRESHOLD)
+                          return Colour.BrightRed
                         return Colour.BrightGreen
                       }
                     )
@@ -513,9 +520,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value, mode) => {
                         switch (mode) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label.guest_toilet_ave
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return (
                               language.ui.main.label.guest_toilet_ave +
                               value.toString()
@@ -530,7 +537,7 @@ function menu(): void {
                     visibility: compute(
                       baseData.local.options.display_mode.store,
                       (value) => {
-                        return value === Options.DisplayMode.PROGRESS_BAR
+                        return value === Data.Options.DisplayMode.PROGRESS_BAR
                           ? "visible"
                           : "none"
                       }
@@ -538,14 +545,15 @@ function menu(): void {
                     percentFilled: compute(
                       baseData.local.guest.guest_toilet_ave.store,
                       (value) => {
-                        return value / 255
+                        return value / GuestData.MAX_TOILET
                       }
                     ),
                     background: Colour.Grey,
                     foreground: compute(
                       baseData.local.guest.guest_toilet_ave.store,
                       (value) => {
-                        if (value / 255 > 0.5) return Colour.BrightRed
+                        if (value > GuestData.TOILET_WARNING_THRESHOLD)
+                          return Colour.BrightRed
                         return Colour.BrightGreen
                       }
                     )
@@ -559,7 +567,6 @@ function menu(): void {
           // Park & Scenario
           groupbox({
             text: language.ui.main.groupbox.park_and_scenario.title,
-            // width: "40%",
             height: "50%",
             content: [
               horizontal([
@@ -719,7 +726,7 @@ function menu(): void {
           // Stalls & Facilities
           groupbox({
             text: language.ui.main.groupbox.stalls_and_facilities.title,
-            height: "25%",
+            height: "10%",
             content: [
               horizontal([
                 label({
@@ -750,10 +757,43 @@ function menu(): void {
                       tr(language.ui.main.label.facilities_count_total, value)
                   )
                 })
-              ]),
-              groupbox({
-                text: language.ui.main.groupbox.stalls_and_facilities.prices,
-                content: []
+              ])
+            ]
+          }),
+          // Finance
+          groupbox({
+            text: language.ui.main.groupbox.finance.title,
+            height: "40%",
+            content: [
+              label({
+                text: compute(
+                  baseData.local.finance.total_income.store,
+                  (value) =>
+                    tr(
+                      language.ui.main.label.finance_total_income,
+                      getCurrencyUnit(value)
+                    )
+                )
+              }),
+              label({
+                text: compute(
+                  baseData.local.finance.total_expenditure.store,
+                  (value) =>
+                    tr(
+                      language.ui.main.label.finance_total_expenditure,
+                      getCurrencyUnit(value)
+                    )
+                )
+              }),
+              label({
+                text: compute(
+                  baseData.local.finance.company_value.store,
+                  (value) =>
+                    tr(
+                      language.ui.main.label.finance_company_value,
+                      getCurrencyUnit(value)
+                    )
+                )
               })
             ]
           })
@@ -762,7 +802,7 @@ function menu(): void {
           // Rides
           groupbox({
             text: language.ui.main.groupbox.rides.title,
-            // width: "30%",
+            height: "50%",
             content: [
               horizontal([
                 label({
@@ -901,6 +941,7 @@ function menu(): void {
           // Options
           groupbox({
             text: language.ui.main.groupbox.options.title,
+            height: "50%",
             content: [
               // Update Status
               horizontal({
@@ -912,13 +953,13 @@ function menu(): void {
                       baseData.local.options.update_status.store,
                       (value) => {
                         switch (value) {
-                          case Options.UpdateStatus.RUNNING:
+                          case Data.Options.UpdateStatus.RUNNING:
                             return language.ui.main.tooltip
                               .options_update_running
-                          case Options.UpdateStatus.MANUAL:
+                          case Data.Options.UpdateStatus.MANUAL:
                             return language.ui.main.tooltip
                               .options_update_manual
-                          case Options.UpdateStatus.PAUSED:
+                          case Data.Options.UpdateStatus.PAUSED:
                             return language.ui.main.tooltip
                               .options_update_paused
                           default:
@@ -931,11 +972,11 @@ function menu(): void {
                       baseData.local.options.update_status.store,
                       (value) => {
                         switch (value) {
-                          case Options.UpdateStatus.RUNNING:
+                          case Data.Options.UpdateStatus.RUNNING:
                             return Sprites.SYNC_RUNNING
-                          case Options.UpdateStatus.MANUAL:
+                          case Data.Options.UpdateStatus.MANUAL:
                             return Sprites.SYNC_MANUAL
-                          case Options.UpdateStatus.PAUSED:
+                          case Data.Options.UpdateStatus.PAUSED:
                             return Sprites.SYNC_PAUSED
                           default:
                             return -1
@@ -944,20 +985,20 @@ function menu(): void {
                     ),
                     onClick: () => {
                       baseData.local.options.update_status.store.set(
-                        Options.UpdateStatus.next(
+                        Data.Options.UpdateStatus.next(
                           baseData.local.options.update_status.store.get()
                         )
                       )
                       switch (
                         baseData.local.options.update_status.store.get()
                       ) {
-                        case Options.UpdateStatus.RUNNING:
+                        case Data.Options.UpdateStatus.RUNNING:
                           interval.resumeAll()
                           break
-                        case Options.UpdateStatus.MANUAL:
+                        case Data.Options.UpdateStatus.MANUAL:
                           interval.pauseManual()
                           break
-                        case Options.UpdateStatus.PAUSED:
+                        case Data.Options.UpdateStatus.PAUSED:
                           interval.pauseAll()
                           break
                         default:
@@ -971,11 +1012,11 @@ function menu(): void {
                       baseData.local.options.update_status.store,
                       (value) => {
                         switch (value) {
-                          case Options.UpdateStatus.RUNNING:
+                          case Data.Options.UpdateStatus.RUNNING:
                             return language.ui.main.label.options_update_running
-                          case Options.UpdateStatus.MANUAL:
+                          case Data.Options.UpdateStatus.MANUAL:
                             return language.ui.main.label.options_update_manual
-                          case Options.UpdateStatus.PAUSED:
+                          case Data.Options.UpdateStatus.PAUSED:
                             return language.ui.main.label.options_update_paused
                           default:
                             return ""
@@ -994,27 +1035,27 @@ function menu(): void {
                     image: Sprites.SYNC_RELOAD,
                     disabled: compute(
                       baseData.local.options.update_status.store,
-                      (value) => value === Options.UpdateStatus.PAUSED
+                      (value) => value === Data.Options.UpdateStatus.PAUSED
                     ),
                     onClick: () => {
                       Data.updateAll()
                       if (
                         baseData.local.options.update_status.store.get() ===
-                        Options.UpdateStatus.MANUAL
+                        Data.Options.UpdateStatus.MANUAL
                       ) {
                         manualIndicatorLit = true
                         baseData.local.options.update_status.store.set(
-                          Options.UpdateStatus.MANUAL // Force an update on subsribers to update the indicator light
+                          Data.Options.UpdateStatus.MANUAL // Force an update on subsribers to update the indicator light
                         )
                         context.setTimeout(() => {
                           manualIndicatorLit = false
                           baseData.local.options.update_status.store.set(
-                            Options.UpdateStatus.MANUAL // Force an update on subsribers to update the indicator light
+                            Data.Options.UpdateStatus.MANUAL // Force an update on subsribers to update the indicator light
                           )
                         }, 1000 * 3)
                       } else if (
                         baseData.local.options.update_status.store.get() ===
-                        Options.UpdateStatus.RUNNING
+                        Data.Options.UpdateStatus.RUNNING
                       ) {
                         // Reset the countdown progress when the sync now button is clicked.
                         interval.syncCounter()
@@ -1037,9 +1078,9 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value) => {
                         switch (value) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return Sprites.SHOW_PROGRESS_BAR
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return Sprites.SHOW_VALUE
                           default:
                             return -1
@@ -1058,10 +1099,10 @@ function menu(): void {
                       baseData.local.options.display_mode.store,
                       (value) => {
                         switch (value) {
-                          case Options.DisplayMode.PROGRESS_BAR:
+                          case Data.Options.DisplayMode.PROGRESS_BAR:
                             return language.ui.main.label
                               .options_display_progress_bar
-                          case Options.DisplayMode.VALUE:
+                          case Data.Options.DisplayMode.VALUE:
                             return language.ui.main.label.options_display_value
                           default:
                             return ""
@@ -1091,4 +1132,4 @@ function menu(): void {
   else main_ui.focus()
 }
 
-export { initMainMenu }
+export { initMainMenu, openMainMenu }